Major logic update
Update logic to copy the randomizer logic.

Hide locations that aren't in the current settings.

Sequence breaks involving locations ([@access_rule]) aren't computed correctly. It should be fixed in a future Emotracker release. It would be possible to correct it manually by creating a bunch of new rules, but I'd rather wait for the release of a fixed version.